+++ trunk/simulations-java/simulations/greenhouse/src/edu/colorado/phet/greenhouse/model/PhotonGlassPaneCollisionModel.java
@@ -41,7 +41,7 @@ public class PhotonGlassPaneCollisionModel {
         double dispersionAngle = Math.PI / 2;
         double theta = Math.random() * dispersionAngle + ( Math.PI * 3 / 2 ) - ( dispersionAngle / 2 );
         theta += Math.random() < 0.5 ? 0 : Math.PI;
-        float vBar = photon.getVelocity().getMagnitude();
+        double vBar = photon.getVelocity().getMagnitude();
         Photon newPhoton = new Photon( photon.getWavelength(), glassPane );
         newPhoton.setVelocity( vBar * (float) Math.cos( theta ),
                                vBar * (float) Math.sin( theta ) );
